Significantly improved perfornce of loading plugins
Without having the need to call self::getPlugin() every time a specific plugin or a plugin type is loaded, it significantly improves performance as we won't be looping through the same plugin array that we are already going through. I tested this with 2000 plugins, the performance going from about ~6 seconds to ~1.4 seconds.
I wrapped the call on an if statement that just checks to see if params is defined, if it is not then we call it, but we never go into this if statement during a normal execution for Joomla, I just left it in there just in case, but if possible I suggest removing it completely.
I will be committing my test cases within the next day or two to the SVN as I am trying to improve the performance of the plugin API.