Fix IS_MAC operating system check
On my OS X 10.8 machine, when I run the code `strtoupper(substr(PHP_OS, 0, 3))`, I get the result of 'DAR', which is short for Darwin. Under the previous revision, the IS_UNIX would be defined as true, and IS_MAC is defined as false, even though I am running the check on a Mac. 

In the platform docs, it says IS_UNIX is only set to true if it is a flavor of UNIX or Linux, but not Mac. This is not true, as Darwin IS Mac, but is labeled as IS_UNIX using the existing code. This fixes that.