Adding $form parameter to JHtmlGrid::sort()
The JS function "Joomla.tableOrdering" allows 4 parameters (order, dir, task and form), but JHtmlGrid::sort() only uses three of them. This commit adds the fourth parameter which allows a developer to choose the name of the form.
This can be handy in case there is more than one form in an output or if the form needs another name for some reason.