Fix to use sessions using file (default behaviour)
Removing the register() function makes file-mode for sessions not working, because the class falls back to JSessionStorage for the register() function, without providing read, write etc. methods.
If you want to use files to save session data, register() needs to exist so that default PHP session-to-file mechanism doesn't stop working.